Add files via upload
The general speed limits in the Philippines in 60 km/h on most highways and 100 km/h on expressways. Your goal here is to compute the speed of automobiles based on distance travelled and time taken to cover this distance.



Note: For this problem, you need to format the output to print the appropriate number of decimal places. You need to use System.out.format (Read: https://docs.oracle.com/javase/tutorial/java/data/numberformat.html)